+++ lib/scripting/napa_script_type.dart
@@ -1,0 +1,5 @@
+/// Script laguages availability
+enum NapaScriptType {
+  /// Lua 5.3 provided by LuaDardo lib.
+  lua,
+}